Fix integer loopback in scanner averages

+++ src/main/java/gregtech/api/metatileentity/BaseMetaTileEntity.java
@@ -57,7 +57,7 @@ public class BaseMetaTileEntity extends BaseTileEntity implements IGregTechTileE
     protected long mStoredEnergy = 0, mStoredSteam = 0;
     protected int mAverageEUInputIndex = 0, mAverageEUOutputIndex = 0;
     protected boolean mReleaseEnergy = false;
-    protected int[] mAverageEUInput = new int[]{0, 0, 0, 0, 0}, mAverageEUOutput = new int[]{0, 0, 0, 0, 0};
+    protected long[] mAverageEUInput = new long[]{0, 0, 0, 0, 0}, mAverageEUOutput = new long[]{0, 0, 0, 0, 0};
     private boolean[] mActiveEUInputs = new boolean[]{false, false, false, false, false, false}, mActiveEUOutputs = new boolean[]{false, false, false, false, false, false};
     private byte[] mSidedRedstone = new byte[]{15, 15, 15, 15, 15, 15};
     private int[] mCoverSides = new int[]{0, 0, 0, 0, 0, 0}, mCoverData = new int[]{0, 0, 0, 0, 0, 0}, mTimeStatistics = new int[GregTech_API.TICKS_FOR_LAG_AVERAGING];
@@ -1701,15 +1701,15 @@ public class BaseMetaTileEntity extends BaseTileEntity implements IGregTechTileE
 
     @Override
     public long getAverageElectricInput() {
-        int rEU = 0;
-        for (int tEU : mAverageEUInput) rEU += tEU;
+        long rEU = 0;
+        for (long tEU : mAverageEUInput) rEU += tEU;
         return rEU / mAverageEUInput.length;
     }
 
     @Override
     public long getAverageElectricOutput() {
-        int rEU = 0;
-        for (int tEU : mAverageEUOutput) rEU += tEU;
+        long rEU = 0;
+        for (long tEU : mAverageEUOutput) rEU += tEU;
         return rEU / mAverageEUOutput.length;
     }
